POC/Streaming_Purchase/Streaming_Purchase_Order_Item/requirements.txt
# Azure Dependencies
azure-eventhub>=5.11.0
azure-core>=1.28.0
azure-identity>=1.14.0
azure-storage-blob>=12.17.0

# Data Processing
pyspark>=3.4.0
delta-spark>=2.4.0
pandas>=2.0.0
numpy>=1.24.0

# Data Validation
pydantic>=2.0.0
jsonschema>=4.19.0

# DQX Framework
databricks-labs-dqx

# Utilities
python-dotenv>=1.0.0
backoff>=2.2.1
structlog>=23.1.0

# Type Hints
typing_extensions>=4.0.0

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-benchmark>=4.0.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0
pytest-rerunfailures>=12.0.0

# Code Quality & Linting
ruff>=0.1.0
black>=23.7.0
isort>=5.12.0
pylint>=2.17.0
mypy>=1.5.0
flake8>=6.1.0

# Security Scanning
bandit[toml]>=1.7.5
safety>=2.3.5
pip-audit>=2.6.0
semgrep>=1.45.0

# Code Complexity Analysis
radon>=6.0.0
mccabe>=0.7.0
xenon>=0.9.0

# Documentation Quality
pydocstyle>=6.3.0
interrogate>=1.5.0
docformatter>=1.7.0

# Coverage & Reporting
coverage[toml]>=7.3.0
coverage-badge>=1.1.0

# Pre-commit Hooks
pre-commit>=3.4.0

# Build & Development Tools
tox>=4.11.0
wheel>=0.41.0
setuptools>=68.0.0

# Additional Security Tools
detect-secrets>=1.4.0
pip-tools>=7.3.0

# Performance Testing
memory-profiler>=0.61.0
psutil>=5.9.0

# Type Checking Extensions
types-requests>=2.31.0
types-PyYAML>=6.0.12
types-setuptools>=68.2.0